make indexing private
