+++ napari/utils/indexing.py
@@ -1,67 +1,11 @@
-from typing import Dict, Tuple
+import warnings
 
-import numpy as np
-import numpy.typing as npt
+from napari.utils._indexing import index_in_slice
 
+__all__ = ['index_in_slice']
 
-def visible_items_in_slice(
-    index: Tuple[npt.NDArray[np.int_], ...], position_in_axes: Dict[int, int]
-) -> npt.NDArray[np.bool_]:
-    """
-    Return a boolean array indicating which items are visible in the current
-    view based on its indices and the current slice.
-
-    Parameters
-    ----------
-    index : tuple of array of int
-        A NumPy fancy indexing expression [1]_.
-    position_in_axes : dict[int, int]
-        A dictionary mapping sliced (non-displayed) axes to a slice position.
-
-    Returns
-    -------
-    visible : array of bool
-        A boolean array indicating which items are visible in the current view.
-    """
-    queries = [
-        index[ax] == position for ax, position in position_in_axes.items()
-    ]
-    return np.logical_and.reduce(queries, axis=0)
-
-
-def index_in_slice(
-    index: Tuple[npt.NDArray[np.int_], ...], position_in_axes: Dict[int, int]
-) -> Tuple[npt.NDArray[np.int_], ...]:
-    """Convert a NumPy fancy indexing expression from data to sliced space.
-
-    Parameters
-    ----------
-    index : tuple of array of int
-        A NumPy fancy indexing expression [1]_.
-    position_in_axes : dict[int, int]
-        A dictionary mapping sliced (non-displayed) axes to a slice position.
-
-    Returns
-    -------
-    sliced_index : tuple of array of int
-        The indexing expression (nD) restricted to the current slice (usually
-        2D or 3D).
-
-    Examples
-    --------
-    >>> index = (np.arange(5), np.full(5, 1), np.arange(4, 9))
-    >>> index_in_slice(index, {0: 3})
-    (array([1]), array([7]))
-    >>> index_in_slice(index, {1: 1, 2: 8})
-    (array([4]),)
-
-    References
-    ----------
-    [1]: https://numpy.org/doc/stable/user/basics.indexing.html#integer-array-indexing
-    """
-    index_in_slice = visible_items_in_slice(index, position_in_axes)
-    return tuple(
-        ix[index_in_slice]
-        for i, ix in enumerate(index)
-        if i not in position_in_axes
-    )
+warnings.warn(
+    "napari.utils.indexing is deprecated since 0.4.19 and will be removed in 0.5.0.",
+    FutureWarning,
+    stacklevel=2,
+)
